THIS_IS_FOR_MAKE_IT_RUN_LOCAL/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
livekit-api>=0.6.0
pydantic>=2.0.0
python-multipart>=0.0.6
livekit-agents[standard]>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-silero>=0.6.0
pinecone-client>=3.0.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
requests>=2.31.0
